adds a higher-level graphics API to SRPG Studio
Just from reading the documentation, it is not obvious how to display an
arbitrary image, let alone how to control its movement or attributes.

Furthermore, there are some things which should be possible, such as
bouncing or moving in a circle, which are unnecessarily verbose to
implement using the native API.

This aims to fill that need.